requirements.txt
apache-airflow[postgres,smtp,web]==2.8.2
flask-appbuilder==4.3.11
alembic>=1.13.1,<2.0
apache-airflow-providers-http>=4.4.1,<5.0
apache-airflow-providers-postgres>=5.7.1,<6.0
apache-airflow-providers-smtp>=1.8.0,<2.0
jupyterlab>=3.0,<5.0
kafka-python>=2.0,<3.0
matplotlib>=3.6,<4.0
numpy
pandas>=1.5,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv
requests>=2.28,<3.0
seaborn
scikit-learn
plotly>=6.1.1,<7.0
kaleido>=1.0.0,<2.0
yfinance